+++ vantage-cli-util/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod model_cli;
 mod table_display;
+pub mod vista_cli;
 
 pub use model_cli::{Mode, ModelFactory, Renderer};
 pub use table_display::{